+++ gulpfile.js
@@ -5,6 +5,7 @@ var merge       = require("merge2");
 var tslint      = require("tslint");
 var gulp        = require("gulp");
 var buffer      = require("gulp-buffer");
+var changed     = require('gulp-changed');
 var sourcemaps  = require("gulp-sourcemaps");
 var tap         = require("gulp-tap");
 var typedoc     = require("gulp-typedoc");
@@ -14,7 +15,7 @@ var gulpTslint  = require("gulp-tslint");
 
 var name = "Mbed Cloud SDK for JavaScript";
 var namespace = "MbedCloudSDK";
-var docsToc = "AccountManagementApi,BootstrapAPI,CertificatesApi,ConnectApi,DeviceDirectoryApi,EnrollmentApi,Subscribe,UpdateApi,ConnectionOptions";
+var docsToc = "AccountManagementApi,BillingApi,BootstrapApi,CertificatesApi,ConnectApi,DeviceDirectoryApi,EnrollmentApi,Subscribe,UpdateApi,ConnectionOptions,SDKError,ListResponse,Paginator";
 
 // Source
 var srcDir = "src";
@@ -90,7 +91,7 @@ gulp.task("doc", function() {
 });
 
 // Build TypeScript source into CommonJS Node modules
-gulp.task("typescript", ["clean"], function() {
+gulp.task("typescript", function() {
     var options = {
         target: "es5",
         types: ["intern"],
@@ -110,6 +111,7 @@ gulp.task("typescript", ["clean"], function() {
 
     return merge([
         gulp.src(srcFiles)
+            .pipe(changed(nodeDir, {extension: '.js'}))
             .pipe(sourcemaps.init())
             .pipe(ts(options))
             .on("error", handleError).js
@@ -118,6 +120,7 @@ gulp.task("typescript", ["clean"], function() {
             }))
             .pipe(gulp.dest(nodeDir)),
         gulp.src(srcFilesOnly)
+            .pipe(changed(typesDir, {extension: '.js'}))
             .pipe(ts(options))
             .on("error", handleError).dts
             .pipe(gulp.dest(typesDir))
@@ -140,7 +143,8 @@ function bundle(srcFiles, destDir, optionsFn) {
             console.log(`Creating ${destDir}/${fileName}`);
 
         file.contents = browserify(file.path, options)
-        .ignore("buffer")
+            .ignore("buffer")
+            .ignore("dotenv")
         .bundle()
         .on("error", handleError);
         file.path = path.join(file.base, fileName);