+++ src/billing/index.ts
@@ -15,15 +15,5 @@
  * limitations under the License.
  */
 
-import { EndpointsBase } from "../common/endpointsBase";
-import { ConnectionOptions } from "../common/interfaces";
-import { PublicAPIApi as EnrollmentApi } from "../_api/enrollment";
-
-export class Endpoints extends EndpointsBase {
-    public readonly enrollment: EnrollmentApi;
-
-    constructor(options?: ConnectionOptions) {
-        super();
-        this.enrollment = new EnrollmentApi(options, this.responseHandler.bind(this));
-    }
-}
+import { BillingApi } from "./billingApi";
+export = BillingApi;