+++ src/bootstrap/bootstrapApi.ts
@@ -18,19 +18,107 @@
 import { apiWrapper, asyncStyle } from "../common/functions";
 import { Endpoints } from "./endpoints";
 import { CallbackFn, ConnectionOptions } from "../common/interfaces";
-import { AddPreSharedKey } from "./types";
+import { AddPreSharedKey, PskListOptions } from "./types";
 import { PreSharedKey } from "./models/preSharedKey";
-import { mapToSDK, mapToSpec, stripToken } from "./models/preSharedKeyAdapter";
+import { mapToSDK, mapToSpec, mapFrom } from "./models/preSharedKeyAdapter";
 import { ApiMetadata } from "../common/apiMetadata";
+import { ListResponse } from "../common/listResponse";
 
 export class BootstrapApi {
     private readonly _endpoints: Endpoints;
 
     /**
+     * The API can be initalized with a .env file in the wroking directory with the following values
+     *
+     * MBED_CLOUD_SDK_API_KEY=<Mbed Cloud Api Key>
+     *
+     * and optionally
+     *
+     * MBED_CLOUD_SDK_HOST=<your host> (defaults to https://api.us-east-1.mbedcloud.com)
+     *
+     * OR
+     * This API is initialized with [ConnectionOptions](../interfaces/connectionoptions.html).
+     *
+     * To create an instance of this API in [Node.js](https://nodejs.org):
+     *
+     * ```JavaScript
+     * var MbedCloudSDK = require("mbed-cloud-sdk");
+     *
+     * var bootstrap = new MbedCloudSDK.BootstrapApi({
+     *     apiKey: "<Mbed Cloud API Key>"
+     * });
+     * ```
+     *
+     * To create an instance of this API in the browser:
+     *
+     * ```html
+     * <script src="<mbed-cloud-sdk>/bundles/update.min.js"></script>
+     *
+     * <script>
+     *     var bootstrap = new MbedCloudSDK.BootstrapApi({
+     *         apiKey: "<Mbed Cloud API Key>"
+     *     });
+     * </script>
+     * ```
      * @param options Connection objects
      */
-    constructor(options: ConnectionOptions) {
+    constructor(options?: ConnectionOptions) {
         this._endpoints = new Endpoints(options);
+    }
+
+     /**
+      * List Psks
+      *
+      * Example:
+      * ```JavaScript
+      * bootstrap.listPsks()
+      * .then(psks => {
+      *     // Utilize psks here
+      * })
+      * .catch(error => {
+      *     console.log(error);
+      * });
+      * ```
+      *
+      * @param options options
+      * @returns Promise of listResponse
+      */
+    public listPsks(options?: PskListOptions): Promise<ListResponse<PreSharedKey>>;
+    /**
+     * List Psks
+     *
+     * Example:
+     * ```JavaScript
+     * bootstrap.listPsks(function(error, psks) {
+     *     if (error) throw error;
+     *     // Utilize psks here
+     * });
+     * ```
+     *
+     * @param options options
+     * @param callback A function that is passed the arguments (error, listResponse)
+     */
+    public listPsks(options?: PskListOptions, callback?: CallbackFn<ListResponse<PreSharedKey>>): void;
+    public listPsks(options?: any, callback?: CallbackFn<ListResponse<PreSharedKey>>): Promise<ListResponse<PreSharedKey>> {
+        options = options || {};
+        if (typeof options === "function") {
+            callback = options;
+            options = {};
+        }
+
+        return apiWrapper(resultsFn => {
+            const { limit, after } = options as PskListOptions;
+            this._endpoints.bootstrap.listPreSharedKeys(limit, after, resultsFn);
+        }, (data, done) => {
+            let keys: Array<PreSharedKey>;
+            if (data && data.data && data.data.length) {
+                keys = data.data.map(key => {
+                    return mapToSDK(key, this);
+                });
+            }
+
+            done(null, new ListResponse(data, keys));
+        }, callback);
     }
 
     /**
@@ -71,7 +159,7 @@ export class BootstrapApi {
         return apiWrapper(resultsFn => {
             this._endpoints.bootstrap.uploadPreSharedKey(mapToSpec(preSharedKey), resultsFn);
         }, (_data, done) => {
-            done(null, stripToken(preSharedKey, this));
+            done(null, mapFrom(preSharedKey, this));
         }, callback);
     }