+++ src/bootstrap/types.ts
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+import { ListOptions } from "../common/interfaces";
+
 /**
  * Pre-Shared Key
  * For more information about such keys, have a look at <a href="https://cloud.mbed.com/docs/latest/connecting/mbed-client-lite-security-considerations.html"/>
@@ -33,4 +35,10 @@ export interface AddPreSharedKey {
      * For example: "4a4a4a4a4a4a4a4a4a4a4a4a4a4a4a4a"
      */
     secretHex: string;
+}
+
+/**
+ * Options to use when listing psks
+ */
+export interface PskListOptions extends ListOptions {
 }