+++ src/common/idGenerator.ts
@@ -15,15 +15,7 @@
  * limitations under the License.
  */
 
-import { EndpointsBase } from "../common/endpointsBase";
-import { ConnectionOptions } from "../common/interfaces";
-import { PublicAPIApi as EnrollmentApi } from "../_api/enrollment";
-
-export class Endpoints extends EndpointsBase {
-    public readonly enrollment: EnrollmentApi;
-
-    constructor(options?: ConnectionOptions) {
-        super();
-        this.enrollment = new EnrollmentApi(options, this.responseHandler.bind(this));
-    }
-}
+// Generates a random URL-friendly identifier string. Note that this is not cryptographically secure.
+export const generateId = (): string => {
+    return `${Date.now()}${Math.floor(Math.random() * 10000)}`;
+};